requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
feedparser>=6.0.10

# LLM
google-generativeai>=0.4.0

# Graph Database
gqlalchemy>=1.5.0
neo4j>=5.0.0

# Embeddings & Vector Search
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4

# NLP / Text
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Commons
datacommons>=1.4.3
datacommons-pandas>=0.0.3

# Web framework (API)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
uuid6>=2024.1.12
tqdm>=4.66.0
aiohttp>=3.9.0
